Combine code from both timelines
Going with the name 'Post' for the timeline component b/c the
alternative, 'Timelime', seems redundant.
